source/flask/api/requirements.txt
flask
numpy
pandas
redis
mysql-connector
paho-mqtt
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client